requirements.txt
frida-tools
pywidevinex